refactor(device-settings): Extract section cards into own files
