+++ app/src/main/java/eu/darken/capod/main/ui/devicesettings/components/SegmentedSettingRow.kt
@@ -1,0 +1,82 @@
+package eu.darken.capod.main.ui.devicesettings.components
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.twotone.Mic
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.SegmentedButton
+import androidx.compose.material3.SegmentedButtonDefaults
+import androidx.compose.material3.SingleChoiceSegmentedButtonRow
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import eu.darken.capod.common.compose.Preview2
+import eu.darken.capod.common.compose.PreviewWrapper
+
+@Composable
+internal fun <T> SegmentedSettingRow(
+    icon: ImageVector,
+    title: String,
+    options: List<Pair<String, T>>,
+    selected: T,
+    onSelected: (T) -> Unit,
+    enabled: Boolean,
+    modifier: Modifier = Modifier,
+    subtitle: String? = null,
+) {
+    Surface(
+        color = MaterialTheme.colorScheme.surfaceContainerLow,
+        shape = RoundedCornerShape(16.dp),
+        modifier = modifier.padding(horizontal = 8.dp, vertical = 4.dp),
+    ) {
+        Column(modifier = Modifier.padding(horizontal = 12.dp, vertical = 12.dp)) {
+            SettingsCompoundHeader(
+                icon = icon,
+                title = title,
+                subtitle = subtitle,
+                enabled = enabled,
+            )
+            Spacer(modifier = Modifier.height(12.dp))
+            SingleChoiceSegmentedButtonRow(modifier = Modifier.fillMaxWidth()) {
+                options.forEachIndexed { index, (label, value) ->
+                    SegmentedButton(
+                        selected = value == selected,
+                        onClick = { if (enabled) onSelected(value) },
+                        enabled = enabled,
+                        shape = SegmentedButtonDefaults.itemShape(index, options.size),
+                        label = {
+                            Text(
+                                text = label,
+                                maxLines = 1,
+                                overflow = TextOverflow.Ellipsis,
+                            )
+                        },
+                    )
+                }
+            }
+        }
+    }
+}
+
+@Preview2
+@Composable
+private fun SegmentedSettingRowPreview() = PreviewWrapper {
+    SegmentedSettingRow(
+        icon = Icons.TwoTone.Mic,
+        title = "Microphone Mode",
+        subtitle = "Which microphone to use during calls",
+        options = listOf("Auto" to "auto", "Left" to "left", "Right" to "right"),
+        selected = "auto",
+        onSelected = {},
+        enabled = true,
+    )
+}
